# Conflicts:
#	ClientApplication/src/main/java/firstFrame/SettingsUser.java
#	ServerApplication/src/main/java/org/example/connection/HttpClient.java
#	demo-battleship/src/main/java/com/example/demo_battleship/controller/GameController.java
#	demo-battleship/src/main/java/com/example/demo_battleship/model/Game.java
#	demo-battleship/src/main/java/com/example/demo_battleship/repository/GameRepository.java

+++ demo-battleship/src/main/java/com/example/demo_battleship/repository/ShipRepository.java
@@ -3,5 +3,5 @@ package com.example.demo_battleship.repository;
 import com.example.demo_battleship.model.Ship;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ShipRepository extends JpaRepository<Ship, Long> {
+public interface ShipRepository extends JpaRepository<Ship, Integer> {
 }
